Add files via upload
This code provides a simple file explorer GUI using Tkinter. Users can click the "Browse" button to select a directory. 

The selected directory is displayed in a label, and the list of files and folders in that directory is displayed in a list box.

GUI Display: 
Browse button: to navigate to a folder in the local directory.